Refactor the `maker_utils` (#193)
