+++ src/roman_datamodels/maker_utils/_base.py
@@ -1,0 +1,2 @@
+NONUM = -999999
+NOSTR = "dummy value"
